Refactor: dto병합 작업 & 사용하지 않는 클래스 삭제
- chat에 있는 dto 병합
- 사용하지 않는 클래스 & 에러 InsufficientTokenException삭제
- Runtime -> BusinessException변경
- dto병합 처리 과정 import 수정

Ref: #195
