--- src/main/java/com/example/baseballprediction/domain/chat/dto/ChatGiftRequestDTO.java
@@ -1,12 +1,0 @@
-package com.example.baseballprediction.domain.chat.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class ChatGiftRequestDTO {
-	private String recipientNickName;
-	private int token;
-	private String comment;
-}
