--- src/main/java/com/example/baseballprediction/domain/chat/dto/ChatLeaveRequest.java
@@ -1,10 +1,0 @@
-package com.example.baseballprediction.domain.chat.dto;
-
-import lombok.Getter;
-
-@Getter
-public class ChatLeaveRequest {
-	
-	 private Long gameId;
-
-}
