+++ src/main/java/com/example/baseballprediction/domain/chat/service/ChatService.java
@@ -23,13 +23,13 @@ import lombok.RequiredArgsConstructor;
 public class ChatService {
 	
 	// 채팅방 ID와 그 채팅방에 참여중인 사용자 세션 ID의 맵핑
-    private Map<String, Set<String>> chatRooms = new ConcurrentHashMap<>();
+    private Map<Long, Set<String>> chatRooms = new ConcurrentHashMap<>();
     //클라이언트가 예기치 못하게게 끊겼을 경우 
     private ConcurrentHashMap<String, String> sessions = new ConcurrentHashMap<>();
     private final GameRepository gameRepository;
     
     // 사용자가 채팅방에 입장할 때 호출됨
-    public void addChatRoom(String sessionId, String gameId) {
+    public void addChatRoom(String sessionId, Long gameId) {
         // 채팅방이 존재하지 않으면 새로 생성
         chatRooms.putIfAbsent(gameId, new HashSet<>());
         // 해당 채팅방에 클라이언트 세션 추가
@@ -37,7 +37,7 @@ public class ChatService {
     }
     
     public void removeMembeSessionChatRoom(String webSocketSessionId, Long gameId) {
-        Set<String> sessions = chatRooms.get(String.valueOf(gameId));
+        Set<String> sessions = chatRooms.get(gameId);
         if (sessions == null) {
         	throw new NotFoundException(ErrorCode.MEMBER_NOT_FOUND);
         }
@@ -45,7 +45,7 @@ public class ChatService {
     }
     
     // 채팅방의 모든 사용자 세션을 종료하는 메서드
-    public void closeChatRoom(String gameId) {
+    public void closeChatRoom(Long gameId) {
     	chatRooms.remove(gameId);
     }
     
@@ -53,14 +53,12 @@ public class ChatService {
         sessions.remove(sessionId);
     }
     
-    //@Scheduled(cron = "0 0/20 2 * * ?", zone = "Asia/Seoul")
-    @Scheduled(fixedDelay = 60000)
+    @Scheduled(cron = "0 0/20 2 * * ?", zone = "Asia/Seoul")
 	public void removeSessionForEndedGames() {
 		List<Long> endedGameIds = gameRepository.findGameIdsByStatus(Status.END);
 		
 		endedGameIds.forEach(gameId -> {
-		    String gameIdString = String.valueOf(gameId);
-		    closeChatRoom(gameIdString);
+		    closeChatRoom(gameId);
 		});
 	}
 }
